changes after review 3
