+++ Core.ApplicationServices/Core.ApplicationServices.csproj
@@ -112,6 +112,7 @@
     <Compile Include="Extensions\IAuthorizationContextExtensions.cs" />
     <Compile Include="Authorization\Policies\GlobalReadAccessPolicy.cs" />
     <Compile Include="Extensions\OptionalValueChangeExtensions.cs" />
+    <Compile Include="Extensions\OrganizationTreeUpdateConsequencesExtensions.cs" />
     <Compile Include="GDPR\DataProcessingRegistrationReadModelService.cs" />
     <Compile Include="GDPR\DataProcessingRegistrationApplicationService.cs" />
     <Compile Include="GDPR\DataProcessingRegistrationOptionsApplicationService.cs" />
