+++ Core.ApplicationServices/Organizations/StsOrganizationSynchronizationService.cs
@@ -5,6 +5,7 @@ using Core.Abstractions.Extensions;
 using Core.Abstractions.Types;
 using Core.ApplicationServices.Authorization;
 using Core.ApplicationServices.Authorization.Permissions;
+using Core.ApplicationServices.Extensions;
 using Core.ApplicationServices.Model.Organizations;
 using Core.DomainModel.Events;
 using Core.DomainModel.Extensions;
@@ -183,16 +184,7 @@ namespace Core.ApplicationServices.Organizations
                             (
                                 importLogResult =>
                                 {
-                                    foreach (var removedLog in importLogResult.RemovedChangeLogs)
-                                    {
-                                        //if a log is part of the AddedChangeLogs (there can only be 5 logs present) it wasn't yet saved to the DB
-                                        if (importLogResult.AddedChangeLogs.Contains(removedLog))
-                                        {
-                                            continue;
-                                        }
-                                        
-                                        _stsChangeLogRepository.Delete(removedLog);
-                                    }
+                                    _stsChangeLogRepository.RemoveRange(importLogResult.RemovedChangeLogs);
                                     _stsChangeLogRepository.Save();
 
                                     return logEntries;
@@ -208,18 +200,6 @@ namespace Core.ApplicationServices.Organizations
         {
             return GetOrganizationWithImportPermission(organizationUuid)
                 .Bind(organization => organization.GetStsOrganizationConnectionEntryLogs(OrganizationUnitOrigin.STS_Organisation, numberOfChangeLogs));
-        }
-
-        public IEnumerable<StsOrganizationConsequenceLog> ConvertConsequencesToConsequenceLogs(OrganizationTreeUpdateConsequences consequences)
-        {
-            var logs = new List<StsOrganizationConsequenceLog>();
-            logs.AddRange(MapAddedOrganizationUnits(consequences));
-            logs.AddRange(MapRenamedOrganizationUnits(consequences));
-            logs.AddRange(MapMovedOrganizationUnits(consequences));
-            logs.AddRange(MapRemovedOrganizationUnits(consequences));
-            logs.AddRange(MapConvertedOrganizationUnits(consequences));
-
-            return logs;
         }
 
         private Result<ExternalOrganizationUnit, OperationError> LoadOrganizationUnits(Organization organization)
@@ -309,89 +289,10 @@ namespace Core.ApplicationServices.Organizations
                 changeLog.UserId = userId;
             }
 
-            changeLog.ConsequenceLogs = ConvertConsequencesToConsequenceLogs(consequences).ToList();
+            changeLog.ConsequenceLogs = consequences.ConvertConsequencesToConsequenceLogs().ToList();
             changeLog.LogTime = DateTime.Now;
             
             return changeLog;
-        }
-
-        private static IEnumerable<StsOrganizationConsequenceLog> MapConvertedOrganizationUnits(OrganizationTreeUpdateConsequences consequences)
-        {
-            return consequences
-                .DeletedExternalUnitsBeingConvertedToNativeUnits
-                .Select(converted => new StsOrganizationConsequenceLog
-                {
-                    Name = converted.organizationUnit.Name,
-                    Type = ConnectionUpdateOrganizationUnitChangeType.Converted,
-                    Uuid = converted.externalOriginUuid,
-                    Description = $"'{converted.organizationUnit.Name}' er slettet i FK Organisation men konverteres til KITOS enhed, da den anvendes aktivt i KITOS."
-                })
-                .ToList();
-        }
-
-        private static IEnumerable<StsOrganizationConsequenceLog> MapRemovedOrganizationUnits(OrganizationTreeUpdateConsequences consequences)
-        {
-            return consequences
-                .DeletedExternalUnitsBeingDeleted
-                .Select(deleted => new StsOrganizationConsequenceLog
-                {
-                    Name = deleted.organizationUnit.Name,
-                    Type = ConnectionUpdateOrganizationUnitChangeType.Deleted,
-                    Uuid = deleted.externalOriginUuid,
-                    Description = $"'{deleted.organizationUnit.Name}' slettes."
-                })
-                .ToList();
-        }
-
-        private static IEnumerable<StsOrganizationConsequenceLog> MapMovedOrganizationUnits(OrganizationTreeUpdateConsequences consequences)
-        {
-            return consequences
-                .OrganizationUnitsBeingMoved
-                .Select(moved =>
-                {
-                    var (movedUnit, oldParent, newParent) = moved;
-                    return new StsOrganizationConsequenceLog
-                    {
-                        Name = movedUnit.Name,
-                        Type = ConnectionUpdateOrganizationUnitChangeType.Moved,
-                        Uuid = movedUnit.ExternalOriginUuid.GetValueOrDefault(),
-                        Description = $"'{movedUnit.Name}' flyttes fra at være underenhed til '{oldParent.Name}' til fremover at være underenhed for {newParent.Name}"
-                    };
-                })
-                .ToList();
-        }
-
-        private static IEnumerable<StsOrganizationConsequenceLog> MapRenamedOrganizationUnits(OrganizationTreeUpdateConsequences consequences)
-        {
-            return consequences
-                .OrganizationUnitsBeingRenamed
-                .Select(renamed =>
-                {
-                    var (affectedUnit, oldName, newName) = renamed;
-                    return new StsOrganizationConsequenceLog
-                    {
-                        Name = oldName,
-                        Type = ConnectionUpdateOrganizationUnitChangeType.Renamed,
-                        Uuid = affectedUnit.ExternalOriginUuid.GetValueOrDefault(),
-                        Description = $"'{oldName}' omdøbes til '{newName}'"
-                    };
-                })
-                .ToList();
-        }
-
-        private static IEnumerable<StsOrganizationConsequenceLog> MapAddedOrganizationUnits(OrganizationTreeUpdateConsequences consequences)
-        {
-            return consequences
-                .AddedExternalOrganizationUnits
-                .Select(added => new StsOrganizationConsequenceLog
-                {
-                    Name = added.unitToAdd.Name,
-                    Type = ConnectionUpdateOrganizationUnitChangeType.Added,
-                    Uuid = added.unitToAdd.Uuid,
-                    Description = $"'{added.unitToAdd.Name}' tilføjes som underenhed til '{added.parent?.Name}'"
-                }
-                )
-                .ToList();
         }
     }
 }
